+++ .gitignore
@@ -13,3 +13,4 @@
 /.idea
 /.nova
 *.secret
+.vscode
